Feat: 메인페이지 구조변경 및 TestList 컴포넌트 추가
